DropdownService & its application on WeeklyBoss.java
